% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_semantic_map.R
\name{plot_semantic_map}
\alias{plot_semantic_map}
\title{Plots all places where a key term is used. Computed over place-keyword matrix.}
\usage{
plot_semantic_map(mat, terms, coords)
}
\description{
Plots all places where a key term is used. Computed over place-keyword matrix.
}
