% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_similarity.R
\name{semantic_similarity}
\alias{semantic_similarity}
\title{Compute similarity over a matrix}
\usage{
semantic_similarity(mat, term, output = "top")
}
\arguments{
\item{mat}{A matrix}

\item{term}{A word (must be in the rownames of mat)}

\item{output}{"top" or "all"}
}
\value{
similarity scores
}
\description{
Compute similarity over a matrix
}

